+++ apps/client/src/widgets/dialog/ShareDialog/libs/types.ts
@@ -1,0 +1,4 @@
+export type ShareDialogProps = {
+  children: React.ReactNode;
+  roomCode: string;
+};
